requirements.txt
fastapi
uvicorn
python-decouple==3.8
httpx
jinja2